Fixed almost all clippy warnings for comms
Eliminated a few unnecessary clones and fixed some code style warnings.

In some cases, warnings are set to allowed to explicitly acknowledge
that the warning is acceptable (and in some cases should be worked on to be removed).

Some lifetime elision are unresolved and cannot be eliminated because they cause
compiler errors.
