+++ comms/dht/src/store_forward/saf_handler/task.rs
@@ -66,6 +66,7 @@ pub struct MessageHandlerTask<S> {
 impl<S> MessageHandlerTask<S>
 where S: Service<DecryptedDhtMessage, Response = (), Error = PipelineError>
 {
+    #[allow(clippy::too_many_arguments)]
     pub fn new(
         config: DhtConfig,
         next_service: S,
