+++ comms/src/message/tag.rs
@@ -24,7 +24,7 @@ use rand::{rngs::OsRng, RngCore};
 use std::fmt;
 
 /// Represents a tag for a message
-#[derive(Clone, Debug, Copy, PartialEq, Eq, Hash)]
+#[derive(Clone, Debug, Copy, PartialEq, Eq, Hash, Default)]
 pub struct MessageTag(u64);
 
 impl MessageTag {
