+++ comms/src/peer_manager/async_peer_manager.rs
@@ -95,6 +95,7 @@ impl AsyncPeerManager {
 
     /// Updates fields for a peer. Any fields set to Some(xx) will be updated. All None
     /// fields will remain the same.
+    #[allow(clippy::too_many_arguments)]
     pub async fn update_peer(
         &self,
         public_key: &CommsPublicKey,
