+++ comms/src/peer_manager/mod.rs
@@ -84,8 +84,8 @@ pub use async_peer_manager::AsyncPeerManager;
 mod peer_id;
 pub use peer_id::PeerId;
 
-mod peer_manager;
-pub use peer_manager::PeerManager;
+mod manager;
+pub use manager::PeerManager;
 
 mod peer_query;
 pub use peer_query::{PeerQuery, PeerQuerySortBy};
