+++ comms/src/peer_manager/peer_storage.rs
@@ -114,6 +114,7 @@ where DS: KeyValueStore<PeerId, Peer>
 
     /// Adds a peer to the routing table of the PeerManager if the peer does not already exist. When a peer already
     /// exist, the stored version will be replaced with the newly provided peer.
+    #[allow(clippy::too_many_arguments)]
     pub fn update_peer(
         &mut self,
         public_key: &CommsPublicKey,
