+++ comms/src/protocol/messaging/test.rs
@@ -20,7 +20,7 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use super::messaging::{
+use super::protocol::{
     MessagingEvent,
     MessagingEventReceiver,
     MessagingProtocol,
