+++ comms/src/tor/control_client/commands/key_value.rs
@@ -20,7 +20,7 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use crate::tor::client::{commands::TorCommand, error::TorClientError, parsers, response::ResponseLine};
+use crate::tor::control_client::{commands::TorCommand, error::TorClientError, parsers, response::ResponseLine};
 use std::{borrow::Cow, marker::PhantomData};
 
 /// The GETCONF command.
