+++ comms/src/tor/control_client/commands/mod.rs
@@ -20,7 +20,7 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use crate::tor::client::response::ResponseLine;
+use crate::tor::control_client::response::ResponseLine;
 
 mod add_onion;
 mod del_onion;
