+++ comms/src/tor/control_client/error.rs
