+++ infrastructure/storage/src/lmdb_store/store.rs
@@ -126,7 +126,7 @@ impl LMDBBuilder {
             builder.set_mapsize(self.db_size_mb * 1024 * 1024)?;
             builder.set_maxdbs(max_dbs)?;
             // Using open::Flags::NOTLS does not compile!?! NOTLS=0x200000
-            let flags = open::Flags::from_bits(0x200000).expect("LMDB open::Flag is correct");
+            let flags = open::Flags::from_bits(0x200_000).expect("LMDB open::Flag is correct");
             builder.open(&self.path, flags, 0o600)?
         };
         let env = Arc::new(env);
