+++ infrastructure/test_utils/src/lib.rs
@@ -8,11 +8,6 @@
 //! - `paths` - Contains utilities which return and create paths which are useful for tests involving files
 //! - `random` - Contains utilities to making generating random values easier
 
-#[macro_use]
-extern crate lazy_static;
-
-pub mod address;
-#[macro_use]
 pub mod enums;
 pub mod futures;
 pub mod paths;
